feat: Add target option to compiler
Target option allows to compile either for node or browser. The
prelude utils like http and such must still be adapted to use the
new macro in the fence.

Big refactor and restructuring of modules as well as cleaning things
and moving typing adt and interface parsing to the canonicalization
phase.

Small additions to package downloader in order to enable installation
of madserver in other projects.